CLC-3460 Acting-as a user updates their first_login_at field
https://jira.ets.berkeley.edu/jira/browse/CLC-3460

Created the acting_as? method in user_api_controller and commented the places I would plan to use it

Used integration testing for the three possible endpoins that would cause this error

/api/my/record_first_login
- you don't want the admin user to record a first log for a "viewed as" user that does not exist in the database

/api/my/status
- you don't want the admin user to record a visit that's wasn't really made by the "viewed as" user

/api/my/opt_out'
- you don't want the admin user to delete an existing "viewed as" user's data rowx

+++ app/controllers/user_api_controller.rb
@@ -23,7 +23,7 @@ class UserApiController < ApplicationController
     if session[:user_id]
       # wrap User::Visit.record_session inside a cache lookup so that we have to write User::Visit records less often.
       self.class.fetch_from_cache session[:user_id] do
-        User::Visit.record session[:user_id]
+        User::Visit.record session[:user_id] unless acting_as?
         true
       end
       status.merge!({
@@ -44,18 +44,22 @@ class UserApiController < ApplicationController
   end
 
   def record_first_login
-    User::Api.new(session[:user_id]).record_first_login
+    User::Api.new(session[:user_id]).record_first_login unless acting_as?
     render :nothing => true, :status => 204
   end
 
   def delete
     if session[:user_id]
-      User::Api.delete(session[:user_id])
+      User::Api.delete(session[:user_id]) unless acting_as?
     end
     render :nothing => true, :status => 204
   end
 
   private
+
+  def acting_as?
+    session[:original_user_id] && (session[:user_id] != session[:original_user_id])
+  end
 
   def acting_as_uid
     if session[:original_user_id] && session[:user_id]
